Resolve "clangのだすwarningをなくす"

Closes #388

See merge request ricos/monolish!273

+++ src/internal/lapack/sytrf/dense_double_sytrf.cpp
@@ -19,7 +19,7 @@ int internal::lapack::sytrf(matrix::Dense<double> &A, std::vector<int> &ipiv) {
   int *ipivd = ipiv.data();
   const char U = 'U';
 
-  if (ipiv.size() != M) {
+  if ((int)ipiv.size() != M) {
     logger.func_out();
     std::runtime_error("lapack::getrf, ipiv size error");
   }
